feat(cache): Add Redis for caching and ETags
Added a Redis cache to keep Fibonacci results for 24 hours. This will speed things up for the numbers that get requested a lot. I also put in ETag support, so if a browser asks for a number it already has, we can just send back a 304 to save some bandwidth.